document_ingestion_service/requirements.txt
# Requirements for document_ingestion_service (deduplicated and cleaned)
fastapi==0.115.9
uvicorn==0.27.1
python-multipart==0.0.9
PyYAML==6.0.1
pdfplumber==0.11.6
pdf2image==1.17.0
camelot-py==0.11.0
pdfminer.six==20250327
PyMuPDF  # PyMuPDF, ensure correct version for your platform
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.58
langchain-ollama==0.3.2
langchain-openai==0.3.16
langchain-text-splitters
chromadb==0.6.3
requests==2.32.3
joblib==1.4.2
python-dotenv==1.0.1
tqdm==4.66.2
opencv-python==4.11.0.86
ghostscript==0.7
scikit-learn==1.6.1
python-docx==1.1.0
docx2pdf
langchain_chroma==0.2.3